requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
APScheduler==3.10.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
stripe==7.10.0
Flask-Limiter==3.5.0
gunicorn==21.2.0
requests==2.31.0